+++ .vscode/settings.json
@@ -1,25 +1,5 @@
 {
   "tailwindCSS.experimental.classRegex": [
     ["tv\\((([^()]*|\\([^()]*\\))*)\\)", "[\"'`]([^\"'`]*).*?[\"'`]"]
-  ],
-  "editor.defaultFormatter": "esbenp.prettier-vscode"
-  // "[javascript]": {
-  //   "editor.defaultFormatter": "biomejs.biome"
-  // },
-  // "[typescript]": {
-  //   "editor.defaultFormatter": "biomejs.biome"
-  // },
-  // "[typescriptreact]": {
-  //   "editor.defaultFormatter": "biomejs.biome"
-  // },
-  // "[javascriptreact]": {
-  //   "editor.defaultFormatter": "biomejs.biome"
-  // },
-  // "[json]": {
-  //   "editor.defaultFormatter": "biomejs.biome"
-  // },
-  // "editor.codeActionsOnSave": {
-  //   "quickfix.biome": "explicit",
-  //   "source.organizeImports.biome": "explicit"
-  // }
+  ]
 }
