+++ package.json
@@ -1,6 +1,7 @@
 {
   "name": "@yutalinux/web-component-lib",
   "version": "1.7.4",
+  "type": "module",
   "scripts": {
     "pr": "./scripts/pullreq.sh",
     "pr:cmd": "gh pr create -f",
@@ -10,11 +11,14 @@
     "build": "vite build && tsc -b",
     "test": "vitest run",
     "test:update": "vitest -u run",
-    "lint": "bun run lint:check",
-    "lint:check": "biome lint",
-    "lint:fix": "biome format --write && biome lint --apply",
-    "format:check": "biome format",
-    "format:fix": "biome format --write"
+    "lint": "eslint",
+    "format": "prettier",
+    "format:fix": "prettier --write",
+    "biome-lint": "bun run lint:check",
+    "biome-lint:check": "biome lint",
+    "biome-lint:fix": "biome format --write && biome lint --apply",
+    "biome-format:check": "biome format",
+    "biome-format:fix": "biome format --write"
   },
   "keywords": [],
   "author": {
@@ -50,6 +54,7 @@
   "devDependencies": {
     "@biomejs/biome": "^1.9.3",
     "@chromatic-com/storybook": "^2.0.2",
+    "@eslint/eslintrc": "^3.1.0",
     "@eslint/js": "^9.12.0",
     "@storybook/addon-essentials": "^8.3.5",
     "@storybook/addon-interactions": "^8.3.5",
@@ -72,6 +77,7 @@
     "eslint-plugin-react-hooks": "5.1.0-rc-fb9a90fa48-20240614",
     "eslint-plugin-react-refresh": "^0.4.12",
     "eslint-plugin-storybook": "^0.9.0",
+    "globals": "^15.11.0",
     "postcss": "^8.4.47",
     "prettier": "^3.3.3",
     "prettier-plugin-tailwindcss": "^0.6.8",
@@ -79,6 +85,7 @@
     "storybook": "^8.3.5",
     "tailwindcss": "^3.4.13",
     "typescript": "^5.6.3",
+    "typescript-eslint": "^8.9.0",
     "vite": "^5.4.8",
     "vite-plugin-dts": "^4.2.4",
     "vite-plugin-lib-inject-css": "^2.1.1",
