+++ src/utils/snapshot.tsx
@@ -1,7 +1,7 @@
 import { type Attributes, type FunctionComponent, createElement } from "react";
 import { renderToString } from "react-dom/server";
 
-export function snapshotElement<T extends {}>(
+export function snapshotElement<T extends object>(
   type: FunctionComponent<T>,
   args: (Attributes & T) | null,
 ) {
