+++ tailwind.config.js
@@ -1,7 +1,5 @@
-const colors = require("tailwindcss/colors");
-
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: ["./src/**/*.{html,js,jsx,ts,tsx,css}"],
   theme: {
     extend: {
