+++ tsconfig.node.json
@@ -24,5 +24,5 @@
       "@/*": ["src/*"]
     }
   },
-  "include": ["vite.config.ts"]
+  "include": ["*.config.ts", "*.config.js"]
 }
